Split Kind into lex/ast variants
We had done this initially, and then unified the types, and now I'm
going back to the original design.

The basic idea here is that the lexer produces coarse-grained 'lexemes',
which are then converted into more detailed types during parsing.
Generally, all lexemes have a corresponding token type (with the
exception of certain special lexemes that represent invalid input data,
such as an unterminated string or an empty hex digit) but it may be a
1-N relationship; that is, a single lexeme may end up being remapped to
various tokens, based on context.

There is one other major motivation for this, and that is `TokenSet`. By
separating these kinds, and making TokenSet use the lexeme type, we
avoid the situation where the user tries to construct a TokenSet that
contains non-lexed tokens, and then is confused when that set doesn't
correctly match as expected.