Arbitrary lookahead parser
With this patch the parser is now capable of arbitrary lookahead. From
some preliminary benchmarking, this comes with a 20% performance cost
during parsing, which I'm not thrilled about. But also: it's not bad,
and it's possible if I do some investigating I will find that this is
something I can improve with a bit of work?