More general pass/fail compile testing
This moves to a new structure for tests, where the `compile-tests`
directory in `test-data` is now used for both expected success and
expected failures. This will make it possible for us to test expected
successes.

This also adds a new env var for testing, FEA_VERBOSE, which will
print failing failure test outputs to the console; and it fixes a bug in
validation that came up in one of the previously not-running test cases.