[GDEF] Handle Attach statements, sort ligature pos
Attach: These had not been supported in fonttools during the first pass at GDEF.
This tweaks the impl somewhat to conform to the behaviour of the
existing tools.

ligature: we were writing these out in the order they appear in the
file, but it looks like they should be sorted.