Compile aalt feature
This was one of the last major outstanding tasks, and it ended up coming
together quite reasonably, considering what a pain in the ass it is:

Basically, we can only compile aalt after compiling other lookups, but
the lookups we generate for aalt go at the beginning of our lookup list;
this means that if we compile aalt, we invalidate all previously
generated gpos layout indices.

These can exist in two places: both in the map of featurekeys: lookups
that is maintained in the CompilationCtx, as well as in any (chain)
contextual sub rules that have been created.

After we compile the aalt lookups, then, we need to iterate through all
of these locations and bump the ids there.

In any case, I think this is working, and this is quite a bit cleaner
than a previous incarnation, so I'm overall happy.