Rework logic for assigning lookups to lang/scripts
This logic is complicated, and it was previously sort of scattered
across compile_ctx; this adds a new type, ActiveFeature, which is
responsible for tracking the various lookups for the various
scripts/languages for the current feature.

This moves more code out of compile_ctx, and also gives us something
nice to write unit tests against.