Ensure no mixing of gsub/gpos lookups
This was inpsired by realizing that it was possible to have a chain rule
that referenced a named lookup of the incorrect type.

To handle this, I now require the user to be explicit when converting a
LookupId to a u16 about whether they are in a GPOS or GSUB context; this
ensures that if lookupids somehow end up in the wrong place, we will at
least crash.