Make ParseTree cheap to clone
We can't display a diagnostic in a useful way without associating it
with a ParseTree (which maps spans in the AST to their originating
source files) but previously the ParseTree type was not cloneable.

This makes ParseTree cloneable, and ensures that it is *cheaply*
cloneable by using Rc.

In addition, I've split out loading sources from storing sources, adding
an internal SourceLoader type to live beside SourceList. This makes it
easier for us to clone SourceList, since we don't need to make
`SourceResovler` cloneable. (This also lets us simplify the API here a
bit, which will be useful for some work to come)