Expose raw tables in Compilation type
This modifies how we present the final result of compilation to the
user. Previously we hid the generated tables, and only presented an API
for retreiving a FontBuilder with all of the tables already added, and
then had some convenience API on top of that for generating a binary
font. This made it annoying to generate the actual tables, forcing the
user to compile, decompile, and finally recompile.

With this patch, the Compilation type has public fields for each table
it that may have been generated, which can be inspected and manipulated
by the user as they see fit.