Check for VarStore in GDEF when deciding to build
we don't bother building a GDEF table if it will be empty, but we had
not updated the is_empty check to consider the possibile existence of
item variation data, and so we were not building GDEF in a situation
where this subtable (and no others) was present.

We should add some integration tests that involve building this table,
but that requires us to have some way to locally wire up the
VariationModel.